refactor: decompose matches/show.tsx into focused components
The match detail page had grown to 1207 lines with inline types,
helpers, hero markup, request card, contact card, and a 156-line
complete-match dialog all in one file. Splitting it into focused
modules cuts the page to 359 lines (-70%) without changing behavior.

- New resources/js/components/match/ subfolder with:
  - types.ts          page-local enriched Team/Match/Event/etc
  - match-hero.tsx    status/teams/score/countdown/action buttons
  - match-requests-card.tsx    incoming requests list
  - opposing-leaders-card.tsx  contact card with WhatsApp helpers
  - complete-match-dialog.tsx  the goal-breakdown dialog
- New lib/match-format.ts holding the status colour/text/date/time
  helpers so they can be reused by other match surfaces later.
- The countdown effect is replaced with the existing
  useMatchCountdown hook which already encoded the same logic.
- show.tsx retains layout, action handlers, and the small cancel
  dialog so the page still owns its top-level state.

No behavior change. bun run types, bun run lint, and composer test
(297 passing) all green.
